backend/requirements.txt
fastapi>=0.68.2
uvicorn>=0.15.0
pydantic>=1.8.2
numpy>=1.24.0
scikit-learn>=0.24.2
pandas>=1.3.3
python-multipart>=0.0.5
sqlalchemy>=1.4.23
python-jose>=3.3.0
passlib>=1.7.4
python-dotenv>=0.19.0
psycopg2-binary>=2.9.1
sentry-sdk>=1.5.0
psutil>=5.8.0
prometheus-client>=0.12.0
python-json-logger>=2.0.7 